Use FluxRefCount based solution to connect to upstream
